Fix database update error
